Moved types to parent folder for use in both functions and client app
