+++ firestore/angular-rewrite/src/app/submit-review-modal/submit-review-modal.component.ts
@@ -19,7 +19,7 @@ import { Firestore, collection, addDoc } from "@angular/fire/firestore";
 import { MatDialogRef, MAT_DIALOG_DATA } from "@angular/material/dialog"
 import { Auth } from '@angular/fire/auth';
 
-import { Rating } from "../review-list/ratings";
+import { Rating } from "../../../types/ratings";
 
 @Component({
   selector: "app-submit-review-modal",
